+++ extra/patterns/diamond.js
@@ -1,17 +1,17 @@
+export default function diamond(){
 let n = 5;
     let string = "";
-    // Upside pyramid
     for (let i = 1; i <= n; i++) {
       for (let j = n; j > i; j--) {
         string += " ";
       }
-       // printing star
+       
       for (let k = 0; k < i; k++) {
         string += "* ";
       }
       string += "\n";
     }
-    // downside pyramid
+    
     for (let i = 0; i <= n-1 ; i++) {
       for (let j = 0; j < i; j++) {
         string += " ";
@@ -23,3 +23,4 @@ let n = 5;
       string += "\n";
     }
     console.log(string);
+  }
