+++ extra/patterns/pyramid.js
@@ -1,5 +1,5 @@
-pyramid(5,'fullpyramid')
-function pyramid(row,type){
+
+export default function pyramid(row,type){
     if(type=='halfpyramid'){
         let n = row; 
     let string = "";
@@ -59,4 +59,3 @@ console.log(string);
 
     }
         }
-
