+++ extra/utils/constants.js
@@ -1,0 +1,8 @@
+export let permissions = {
+    'getUsers': {
+        all: ['head-trainer'],
+        read: ['trainee', 'trainer'],
+        write: ['trainer'],
+        delete: [],
+    }
+};
