+++ extra/utils/helpers.js
@@ -1,0 +1,9 @@
+export function ValidateEmail(email) {
+	var emailformat = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+	if (email.match(emailformat)) {
+    	return true;
+    }
+    else {
+    return false;
+}
+}
