+++ extra/utils/index.js
@@ -1,2 +1,2 @@
-export{utils as permission} from './permissions.js';
-export{utils as validation} from './validation.js';
+export {default as hasPermission} from './permissions';
+export {default as validateuser} from './validation';
