+++ extra/utils/validation.js
@@ -1,10 +1,9 @@
-const users = ["abcd123@mail.com","ankit@mail.com", "patidar123@mail.com", "ankit", "patidar", "ab123"];
-valisateuser(users);
-function valisateuser(users){
+import {ValidateEmail} from './helpers';
+export default function validateuser(users){
     const valid =[];
     const invalid =[];
     let res = [];
-    userlen = users.length;
+    let userlen = users.length;
     for(var i=0; i<userlen;i++){
         res = ValidateEmail(users[i]);
         if (res == "true"){
@@ -20,13 +19,4 @@ function valisateuser(users){
     var invlaid_user = invalid.length;
     console.log(`number of valid user= ${valid_user}` );
     console.log(`number of invalid user= ${invlaid_user}` );
-
-}
-
-function ValidateEmail(email) {
-	var emailformat = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-	if (email.match(emailformat)) {
-    	return "true";
-    };
-    return (false);
 }
